Separate account repository responsibilities #532
