+++ code/FinanceManager.Infrastructure/Repositories/Account/Entry/BondEntryRepository.cs
@@ -9,6 +9,7 @@ namespace FinanceManager.Infrastructure.Repositories.Account.Entry;
 
 public class BondEntryRepository(AppDbContext context) : IBondAccountEntryRepository<BondAccountEntry>
 {
+    private readonly BondEntryValueCalculator _valueCalculator = new(context);
     public async Task<bool> Add(BondAccountEntry entry, bool recalculate)
     {
         // Don't use entry.Value as it may be a placeholder (-1)
@@ -342,105 +343,7 @@ public class BondEntryRepository(AppDbContext context) : IBondAccountEntryReposi
     }
 
     private async Task RecalculateValues(int accountId, DateTime startDate)
-    {
-        if (!context.Database.IsRelational())
-        {
-            await RecalculateValuesInMemory(accountId, startDate);
-            return;
-        }
-
-        if (context.Database.ProviderName?.StartsWith("Npgsql") == true)
-        {
-            // DISTINCT ON picks the most-recent entry per BondDetailsId before startDate as the anchor.
-            await context.Database.ExecuteSqlAsync($"""
-                WITH anchors AS (
-                    SELECT DISTINCT ON ("BondDetailsId") "BondDetailsId", "Value" AS "AnchorValue"
-                    FROM "BondEntries"
-                    WHERE "AccountId" = {accountId} AND "PostingDate" < {startDate}
-                    ORDER BY "BondDetailsId", "PostingDate" DESC, "EntryId" DESC
-                ),
-                running AS (
-                    SELECT e."EntryId",
-                           COALESCE(a."AnchorValue", 0) + SUM(e."ValueChange") OVER (
-                               PARTITION BY e."BondDetailsId"
-                               ORDER BY e."PostingDate", e."EntryId"
-                               ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW
-                           ) AS "NewValue"
-                    FROM "BondEntries" e
-                    LEFT JOIN anchors a ON a."BondDetailsId" = e."BondDetailsId"
-                    WHERE e."AccountId" = {accountId} AND e."PostingDate" >= {startDate}
-                )
-                UPDATE "BondEntries" AS e
-                SET "Value" = r."NewValue"
-                FROM running AS r
-                WHERE e."EntryId" = r."EntryId"
-                """);
-        }
-        else
-        {
-            await context.Database.ExecuteSqlAsync($"""
-                WITH anchors_raw AS (
-                    SELECT BondDetailsId, Value AS AnchorValue,
-                           ROW_NUMBER() OVER (PARTITION BY BondDetailsId ORDER BY PostingDate DESC, EntryId DESC) AS rn
-                    FROM BondEntries
-                    WHERE AccountId = {accountId} AND PostingDate < {startDate}
-                ),
-                anchors AS (
-                    SELECT BondDetailsId, AnchorValue FROM anchors_raw WHERE rn = 1
-                ),
-                running AS (
-                    SELECT e.EntryId,
-                           COALESCE(a.AnchorValue, 0) + SUM(e.ValueChange) OVER (
-                               PARTITION BY e.BondDetailsId
-                               ORDER BY e.PostingDate, e.EntryId
-                               ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW
-                           ) AS NewValue
-                    FROM BondEntries e
-                    LEFT JOIN anchors a ON a.BondDetailsId = e.BondDetailsId
-                    WHERE e.AccountId = {accountId} AND e.PostingDate >= {startDate}
-                )
-                UPDATE e
-                SET Value = r.NewValue
-                FROM BondEntries e
-                INNER JOIN running r ON e.EntryId = r.EntryId
-                """);
-        }
-    }
-
-    private async Task RecalculateValuesInMemory(int accountId, DateTime startDate)
-    {
-        var entries = await context.BondEntries
-            .Where(e => e.AccountId == accountId && e.PostingDate >= startDate)
-            .OrderBy(e => e.PostingDate)
-            .ThenBy(e => e.EntryId)
-            .ToListAsync();
-
-        if (entries.Count == 0) return;
-
-        var bondIds = entries.Select(e => e.BondDetailsId).Distinct().ToList();
-
-        // Single query for all bond anchors instead of one per BondDetailsId group.
-        var anchors = (await context.BondEntries
-                .AsNoTracking()
-                .Where(e => e.AccountId == accountId && bondIds.Contains(e.BondDetailsId) && e.PostingDate < startDate)
-                .OrderByDescending(e => e.PostingDate)
-                .ThenByDescending(e => e.EntryId)
-                .ToListAsync())
-            .GroupBy(e => e.BondDetailsId)
-            .ToDictionary(g => g.Key, g => g.First().Value);
-
-        foreach (var bondGroup in entries.GroupBy(e => e.BondDetailsId))
-        {
-            decimal running = anchors.TryGetValue(bondGroup.Key, out var anchor) ? anchor : 0m;
-            foreach (var e in bondGroup.OrderBy(e => e.PostingDate).ThenBy(e => e.EntryId))
-            {
-                running += e.ValueChange;
-                e.Value = running;
-            }
-        }
-
-        await context.SaveChangesAsync();
-    }
+        => await _valueCalculator.Recalculate(accountId, startDate);
 
     public async Task<bool> AddLabel(int entryId, int labelId)
     {
