+++ code/FinanceManager.Infrastructure/Repositories/Account/Entry/CurrencyEntryRepository.cs
@@ -8,6 +8,7 @@ namespace FinanceManager.Infrastructure.Repositories.Account.Entry;
 
 public class CurrencyEntryRepository(AppDbContext context) : IAccountEntryRepository<CurrencyAccountEntry>
 {
+    private readonly CurrencyEntryValueCalculator _valueCalculator = new(context);
     public async Task<bool> Add(CurrencyAccountEntry entry, bool recalculate)
     {
         CurrencyAccountEntry newAccountEntry = new(entry.AccountId, 0, entry.PostingDate, entry.Value, entry.ValueChange)
@@ -384,77 +385,7 @@ public class CurrencyEntryRepository(AppDbContext context) : IAccountEntryReposi
     }
 
     private async Task RecalculateValues(int accountId, DateTime startDate)
-    {
-        var anchor = await context.CurrencyEntries
-            .AsNoTracking()
-            .Where(e => e.AccountId == accountId && e.PostingDate < startDate)
-            .OrderByDescending(e => e.PostingDate)
-            .ThenByDescending(e => e.EntryId)
-            .Select(e => (decimal?)e.Value)
-            .FirstOrDefaultAsync();
-
-        if (!context.Database.IsRelational())
-        {
-            await RecalculateValuesInMemory(accountId, startDate, anchor);
-            return;
-        }
-
-        if (context.Database.ProviderName?.StartsWith("Npgsql") == true)
-        {
-            await context.Database.ExecuteSqlAsync($"""
-                WITH running AS (
-                    SELECT "EntryId",
-                           {anchor ?? 0m} + SUM("ValueChange") OVER (
-                               ORDER BY "PostingDate", "EntryId"
-                               ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW
-                           ) AS "NewValue"
-                    FROM "CurrencyEntries"
-                    WHERE "AccountId" = {accountId}
-                      AND "PostingDate" >= {startDate}
-                )
-                UPDATE "CurrencyEntries" AS e
-                SET "Value" = r."NewValue"
-                FROM running AS r
-                WHERE e."EntryId" = r."EntryId"
-                """);
-        }
-        else
-        {
-            await context.Database.ExecuteSqlAsync($"""
-                WITH running AS (
-                    SELECT EntryId,
-                           {anchor ?? 0m} + SUM(ValueChange) OVER (
-                               ORDER BY PostingDate, EntryId
-                               ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW
-                           ) AS NewValue
-                    FROM CurrencyEntries
-                    WHERE AccountId = {accountId}
-                      AND PostingDate >= {startDate}
-                )
-                UPDATE e
-                SET Value = r.NewValue
-                FROM CurrencyEntries AS e
-                INNER JOIN running AS r ON e.EntryId = r.EntryId
-                """);
-        }
-    }
-
-    private async Task RecalculateValuesInMemory(int accountId, DateTime startDate, decimal? anchor)
-    {
-        var entries = await context.CurrencyEntries
-            .Where(e => e.AccountId == accountId && e.PostingDate >= startDate)
-            .OrderBy(e => e.PostingDate)
-            .ThenBy(e => e.EntryId)
-            .ToListAsync();
-
-        decimal running = anchor ?? 0m;
-        foreach (var e in entries)
-        {
-            running += e.ValueChange;
-            e.Value = running;
-        }
-        await context.SaveChangesAsync();
-    }
+        => await _valueCalculator.Recalculate(accountId, startDate);
 
     public async Task<IReadOnlyList<CurrencyAccountEntry>> GetByIds(IReadOnlyCollection<int> entryIds, CancellationToken cancellationToken = default)
     {
